FEAT: The pattern API can now handle being being given a device id with null profiles as evidence.
When a null profile (a profile with an id of zero) is given, instead of thowning an exception about it being invalid, the handling is defered to the point where a value is fetched. When getting a value for a null profile, hasValue will be false with the reason being reported as a null profile.