TEST: Engine tests now take a list of data files, and some validation tests have been fixed.
By taking a list of data files, an Enterprise file can be used for testing, but if one does not exist, the Lite file can still be used to run the tests.
Running with the extended properties in the Enterprise data file uncovered a bug in the tests where values were not being dereferenced.