Merged PR 6159: EXAMPLE: Added User Agent Client Hint (UACH) headers to the Getting Started e...
EXAMPLE: Added User Agent Client Hint (UACH) headers to the Getting Started example for CPP.

Amended the string of requirement properties to include the SetHeaderPlatfromAccept-CH property to inform device detection that UACH for Platform will be required.

Added PlatformVersion and PlatformName to the example required properties to demonstrate how the values returned change when UACH header values are used.